feat: Implement search endpoints for organizations and providers, and add lab unit configuration in their respective domains
