+++ src/main/java/org/openelisglobal/provider/controller/rest/ProviderRestController.java
@@ -1,6 +1,12 @@
 package org.openelisglobal.provider.controller.rest;
 
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.UUID;
+import org.apache.commons.validator.GenericValidator;
+import org.openelisglobal.common.log.LogEvent;
 import org.openelisglobal.person.service.PersonService;
 import org.openelisglobal.person.valueholder.Person;
 import org.openelisglobal.provider.service.ProviderService;
@@ -53,6 +59,80 @@ public class ProviderRestController {
             return ResponseEntity.ok(updatedProvider);
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Error processing request.");
+        }
+    }
+
+    @GetMapping(value = "/provider/search", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Map<String, Object>> searchProviders(@RequestParam(required = false) String search,
+            @RequestParam(required = false) String phone, @RequestParam(defaultValue = "1") int page,
+            @RequestParam(defaultValue = "20") int pageSize) {
+
+        try {
+            List<Provider> providers;
+            int totalCount;
+            int startRecNo = ((page - 1) * pageSize) + 1;
+
+            if (!GenericValidator.isBlankOrNull(phone)) {
+                providers = providerService.getPagesOfSearchedProvidersByPhone(startRecNo, phone);
+                totalCount = providerService.getTotalSearchedProviderCountByPhone(phone);
+            } else if (!GenericValidator.isBlankOrNull(search)) {
+                providers = providerService.getPagesOfSearchedProviders(startRecNo, search);
+                totalCount = providerService.getTotalSearchedProviderCount(search);
+            } else {
+                providers = providerService.getPageOfProviders(startRecNo);
+                totalCount = (int) providerService.getCount();
+            }
+
+            List<Map<String, Object>> results = new ArrayList<>();
+            for (Provider provider : providers) {
+                Map<String, Object> providerData = new HashMap<>();
+                providerData.put("id", provider.getId());
+
+                if (provider.getPerson() != null) {
+                    personService.getData(provider.getPerson());
+                    providerData.put("personId", provider.getPerson().getId());
+                    providerData.put("firstName", provider.getPerson().getFirstName());
+                    providerData.put("lastName", provider.getPerson().getLastName());
+
+                    String fullName = "";
+                    if (provider.getPerson().getLastName() != null) {
+                        fullName = provider.getPerson().getLastName();
+                    }
+                    if (provider.getPerson().getFirstName() != null) {
+                        if (!fullName.isEmpty()) {
+                            fullName += ", ";
+                        }
+                        fullName += provider.getPerson().getFirstName();
+                    }
+                    providerData.put("name", fullName);
+
+                    String providerPhone = provider.getPerson().getPrimaryPhone();
+                    if (GenericValidator.isBlankOrNull(providerPhone)) {
+                        providerPhone = provider.getPerson().getWorkPhone();
+                    }
+                    providerData.put("phone", providerPhone);
+                    providerData.put("fax", provider.getPerson().getFax());
+                    providerData.put("email", provider.getPerson().getEmail());
+                }
+
+                providerData.put("externalId", provider.getExternalId());
+                providerData.put("isActive", "Y".equals(provider.getActive()));
+
+                results.add(providerData);
+            }
+
+            Map<String, Object> response = new HashMap<>();
+            response.put("providers", results);
+            response.put("totalCount", totalCount);
+            response.put("page", page);
+            response.put("pageSize", pageSize);
+
+            return ResponseEntity.ok(response);
+
+        } catch (Exception e) {
+            LogEvent.logError(this.getClass().getName(), "searchProviders",
+                    "Error searching providers: " + e.getMessage());
+            return ResponseEntity.internalServerError().build();
         }
     }
 }
